+++ src/ServiceInstancesMonitor.js
@@ -1,15 +1,13 @@
-'use strict';
-
 const EventEmitter = require('events');
 const _ = require('lodash');
 const instancesFactory = require('./Factory');
 const ServiceInstances = require('./ServiceInstances');
 const WatchError = require('./Error').WatchError;
-const WatchTimeoutError = require('./Error').WatchTimeoutError;
 const AlreadyInitializedError = require('./Error').AlreadyInitializedError;
 
 const DEFAULT_TIMEOUT_MSEC = 5000;
 const HEALTH_FALLBACK_INTERVAL_MSEC = 1000;
+const DEFAULT_RETRY_START_SERVICE_TIMEOUT_MSEC = 1000;
 
 /**
  * Single node data
@@ -28,6 +26,9 @@ const HEALTH_FALLBACK_INTERVAL_MSEC = 1000;
 /**
  * @emits ServiceInstancesMonitor#initialized
  * @emits ServiceInstancesMonitor#changed
+ * @emits ServiceInstancesMonitor#error
+ * @emits ServiceInstancesMonitor#healthy
+ * @emits ServiceInstancesMonitor#unhealthy
  */
 class ServiceInstancesMonitor extends EventEmitter {
 
@@ -35,7 +36,7 @@ class ServiceInstancesMonitor extends EventEmitter {
      * @param {Object} options
      * @param {string} options.serviceName -  name of service in consul to monitor
      * @param {string} options.checkNameWithStatus
-     * @param {string} [options.timeoutMsec=5000] - connection timeout to consul
+     * @param {number} [options.timeoutMsec=5000] - connection timeout to consul
      * @param {Consul} consul
      * @param {Object} extractors
      * @throws {TypeError} On invalid options format
@@ -96,6 +97,7 @@ class ServiceInstancesMonitor extends EventEmitter {
         this._onWatcherChange = this._onWatcherChange.bind(this);
         this._onWatcherError = this._onWatcherError.bind(this);
         this._onWatcherEnd = this._onWatcherEnd.bind(this);
+        this._retryStartService = this._retryStartService.bind(this);
 
         this._serviceInstances = new ServiceInstances();
         this._watchAnyNodeChange = null;
@@ -103,6 +105,7 @@ class ServiceInstancesMonitor extends EventEmitter {
         this._setUninitialized();
 
         this._fallbackToWatchHealthyInterval = null;
+        this._retryTimer = null;
     }
 
     isWatchHealthy() {
@@ -138,18 +141,18 @@ class ServiceInstancesMonitor extends EventEmitter {
     }
 
     /**
-     * Starts service and resolves promise with initial list of nodes that provide the service.
+     * Starts service and resolves promise with initial list of nodes that provide the service, "change" or "error"
+     * events will not be emited
      *
      * Listens for changes after successful resolve.
      *
      * Promise will be rejected with:
      *   `AlreadyInitializedError` if service is already started.
-     *   `WatchTimeoutError` if either initial data nor error received for 5000 msec
      *   `WatchError` on error from `consul` underlying method
      *
      * Rejection of promise means that watcher was stopped and no retries will be done.
      *
-     * @returns {Promise<Array,AlreadyInitializedError|WatchError|WatchTimeoutError>}
+     * @returns {Promise<ServiceInstances,AlreadyInitializedError|WatchError>}
      * @public
      */
     startService() {
@@ -177,7 +180,6 @@ class ServiceInstancesMonitor extends EventEmitter {
      *
      * Promise will be rejected with:
      *   `AlreadyInitializedError` if service is already started.
-     *   `WatchTimeoutError` if either initial data nor error received for 5000 msec
      *   `WatchError` on error from `consul` underlying method
      *
      * Rejection of promise means that watcher was stopped and no retries will be done.
@@ -186,6 +188,11 @@ class ServiceInstancesMonitor extends EventEmitter {
      * @public
      */
     stopService() {
+        if (this._retryTimer !== null) {
+            clearTimeout(this._retryTimer);
+            this._retryTimer = null;
+        }
+
         if (!this._isWatcherRegistered()) {
             return this;
         }
@@ -197,6 +204,7 @@ class ServiceInstancesMonitor extends EventEmitter {
         this._unsetFallbackToWatchHealthy();
         this._setUninitialized();
         this._setWatchUnealthy();
+
         return this;
     }
 
@@ -208,18 +216,17 @@ class ServiceInstancesMonitor extends EventEmitter {
      *
      * Promise will be rejected with:
      *   `AlreadyInitializedError` if another `consul.watch` execution is found.
-     *   `WatchTimeoutError` if either initial data nor error received for 5000 msec
      *   `WatchError` on error from `consul` underlying method
      *
      * Rejection of promise means that watch was stopped and `this._watchAnyNodeChange` was cleared.
      *
-     * @returns {Promise<Array,AlreadyInitializedError|WatchError|WatchTimeoutError>}
+     * @returns {Promise<ServiceInstances,AlreadyInitializedError|WatchError>}
      * @private
      */
     _registerWatcherAndWaitForInitialNodes() {
         return new Promise((resolve, reject) => {
             if (this._watchAnyNodeChange !== null) {
-                reject(new AlreadyInitializedError('Another `consul.watch` execution is found'));
+                return reject(new AlreadyInitializedError('Another `consul.watch` execution is found'));
             }
 
             this._watchAnyNodeChange = this._consul.watch({
@@ -227,12 +234,12 @@ class ServiceInstancesMonitor extends EventEmitter {
                 options: {
                     service: this._serviceName,
                     wait: '60s',
+                    timeout: this._timeoutMsec
                 },
             });
 
             const firstChange = (data) => {
                 this._watchAnyNodeChange.removeListener('error', firstError);
-                clearTimeout(timerId);
 
                 const {instances, errors} = instancesFactory.buildServiceInstances(
                     data,
@@ -251,17 +258,8 @@ class ServiceInstancesMonitor extends EventEmitter {
                 this._watchAnyNodeChange.removeListener('change', firstChange);
                 this._watchAnyNodeChange.end();
                 this._watchAnyNodeChange = null;
-                clearTimeout(timerId);
                 reject(new WatchError(err.message, {err}));
             };
-
-            const timerId = setTimeout(() => {
-                this._watchAnyNodeChange.removeListener('error', firstError);
-                this._watchAnyNodeChange.removeListener('change', firstChange);
-                this._watchAnyNodeChange.end();
-                this._watchAnyNodeChange = null;
-                reject(new WatchTimeoutError('Initial consul watch request was timed out'));
-            }, this._timeoutMsec);
 
             this._watchAnyNodeChange.once('change', firstChange);
             this._watchAnyNodeChange.once('error', firstError);
@@ -269,18 +267,20 @@ class ServiceInstancesMonitor extends EventEmitter {
     }
 
     /**
-     * This method receives list of healthy nodes sent by `consul.watch` in `consul` format. Performs
+     * This method receives list of a valid nodes sent by `consul.watch` in `consul` format. Performs
      * validation of response format.
      *
      * If service was unhealthy, it becomes healthy.
      *
      * @param {Array} data - list of healthy nodes after some changes
-     * @emits ServiceInstancesMonitor#changed actual array of healthy nodes
+     * @emits ServiceInstancesMonitor#changed actual array of a valid nodes
      * @private
      */
     _onWatcherChange(data) {
+        let isHealthyStateChanged = false;
         if (!this.isWatchHealthy()) {
             this._setWatchHealthy();
+            isHealthyStateChanged = true;
         }
 
         const {instances, errors} = instancesFactory.buildServiceInstances(
@@ -290,6 +290,9 @@ class ServiceInstancesMonitor extends EventEmitter {
         );
 
         this._serviceInstances = instances;
+        if (isHealthyStateChanged) {
+            this.emit('healthy');
+        }
         this.emit('changed', instances);
 
         if (!_.isEmpty(errors)) {
@@ -302,6 +305,7 @@ class ServiceInstancesMonitor extends EventEmitter {
 
         if (this.isWatchHealthy()) {
             this._setWatchUnealthy();
+            this.emit('unhealthy');
         }
 
         this._setFallbackToWatchHealthy();
@@ -309,11 +313,18 @@ class ServiceInstancesMonitor extends EventEmitter {
         this.emit('error', new WatchError(err.message, {err}));
     }
 
-    _onWatcherEnd() {
+    async _onWatcherEnd() {
+        this._unsetFallbackToWatchHealthy();
         this._setUninitialized();
-        this._setWatchUnealthy();
+
+        if (this.isWatchHealthy()) {
+            this._setWatchUnealthy();
+            this.emit('unhealthy');
+        }
+
+        this._watchAnyNodeChange.removeAllListeners();
         this._watchAnyNodeChange = null;
-        this.emit('emergencyStop');
+        await this._retryStartService();
     }
 
     _emitFactoryErrors(errors) {
@@ -330,11 +341,8 @@ class ServiceInstancesMonitor extends EventEmitter {
         const initialUpdateTime = this._watchAnyNodeChange.updateTime();
 
         this._fallbackToWatchHealthyInterval = setInterval(() => {
-            const isWatcherRunning = this._isWatcherRegistered() && this._watchAnyNodeChange.isRunning();
-
-            if (!isWatcherRunning || this.isWatchHealthy()) {
-
-                // watcher is currently ends or becomes `healthy`, unset fallback interval',
+            if (this.isWatchHealthy()) {
+                // watcher is currently becomes `healthy`, unset fallback interval',
                 this._unsetFallbackToWatchHealthy();
 
                 return;
@@ -346,6 +354,7 @@ class ServiceInstancesMonitor extends EventEmitter {
                 this._unsetFallbackToWatchHealthy();
 
                 this._setWatchHealthy();
+                this.emit('healthy');
             }
 
         }, HEALTH_FALLBACK_INTERVAL_MSEC);
@@ -357,6 +366,20 @@ class ServiceInstancesMonitor extends EventEmitter {
         this._fallbackToWatchHealthyInterval = null;
     }
 
+
+    async _retryStartService() {
+        try {
+            const serviceInstances = await this.startService();
+            this._serviceInstances = serviceInstances;
+
+            this.emit('healthy');
+            this.emit('changed', serviceInstances);
+        } catch (err) {
+            setImmediate(() => this.emit('error', err));
+
+            this._retryTimer = setTimeout(this._retryStartService, DEFAULT_RETRY_START_SERVICE_TIMEOUT_MSEC);
+        }
+    }
 }
 
 module.exports = ServiceInstancesMonitor;