+++ CHANGELOG.md
@@ -2,8 +2,18 @@
 
 ### 2.0.0
 
-- [#19681](https://redmine.hwtool.net/issues/19681) Extending consul-service-health-monitor to extract custom data 
-from output
+UPDATES:
+
+- [#20123](https://redmine.hwtool.net/issues/20123) Add to "consul-service-health-monitor" module functionality of an auto-reconnect to Сonsul
+- [#19681](https://redmine.hwtool.net/issues/19681) Extending consul-service-health-monitor to extract custom data from output
+- support auto-reconnect to Consul 
+- error `WatchTimeoutError` is not supported already
+- removed `emergencyStop` event from `ServiceInstancesMonitor`
+- removed `ServiceInstanceStatus` class and added `ServiceInstanceInfo`
+- removed methods `addOnMaintenance`, `addOverloaded`, `getOverloaded` and `getOnMaintenance` from `ServiceInstances` class.
+- removed method `getStatus` from `ServiceInstance` and added `getInfo`
+- added `extractors` argument to `ServiceInstancesMonitor::constructor`
+- added `healthy` and `unhealthy` events to `ServiceInstancesMonitor`
 
 ### 1.4.0